Refactor eslint impl into its own file
